PS-5008: Memory leak in sync_latch_meta_init() after mysqld shutdown detected by ASan
Free all the resources acquired by InnoDB (mutexes, events, memory) in
case of a failed start (errors in
storage/innobase/srv/srv0start.cc::innobase_start_or_create_for_mysql())

Removed silencing leaks in tests

Added missing `innodb_init_abort()` in `innodb_init()`. It caused leaks

Added missing `recv_recovery_from_checkpoint_finish()` in `srv_start()`. It caused leaks

+++ storage/innobase/lock/lock0lock.cc
@@ -383,6 +383,8 @@ void lock_sys_close(void) {
     lock_latest_err_file = nullptr;
   }
 
+  if (!lock_sys) return;
+
   ut::delete_(lock_sys->rec_hash);
   ut::delete_(lock_sys->prdt_hash);
   ut::delete_(lock_sys->prdt_page_hash);
